% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varEqChol.R
\name{compareGraph}
\alias{compareGraph}
\title{Compares the estimated and true adjacency matrices}
\usage{
compareGraph(estAdj, trueAdj)
}
\arguments{
\item{estAdj}{- estimated adjacency matrix}

\item{trueAdj}{- true adjacency matrix}
}
\value{
True Positive, False Positive and True Discovery rates
}
\description{
Compares the estimated and true adjacency matrices
}
