% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varEqChol.R
\name{eqvarDAG_ch}
\alias{eqvarDAG_ch}
\title{Returns ordering from the Cholesky factor}
\usage{
eqvarDAG_ch(X, type = c("low", "high"), crit = c("CV", "eBIC"),
  gamma = 0.5, refine = FALSE, lamlist = NULL, nlam = 30,
  flmin = 0.01, mtd = "dlasso", alpha = 0.05, threshold = 0.1,
  FCD = TRUE, precmtd = "sqrtlasso")
}
\arguments{
\item{X}{- n x p data}

\item{type}{- type of dimension}

\item{crit}{- criterion for penalty parameter selection}

\item{gamma}{- parameter for eBIC. Active if crit = "eBIC"}

\item{refine}{- if TRUE, refines edges after CSCS algorithm}

\item{lamlist}{- A list of non-negative tuning patameters for CV}

\item{nlam}{- if lamlist = NULL, create list with length equal to nlam}

\item{flmin}{- If lamlist is not provided, create a lamlist with ratio of the smallest and largest}

\item{mtd}{- parameters for refine}

\item{alpha}{- parameters for refine}

\item{threshold}{- threshold value}

\item{FCD}{- parameters for refine}

\item{precmtd}{- parameters for refine}

\item{flmin}{- in the list equal to flmin. Default is 0.01.}
}
\value{
- ordering, lower triangular Cholesky factor and adjacency matrix
}
\description{
Returns ordering from the Cholesky factor
}
