% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varEqChol.R
\name{EqVarDAG_HD}
\alias{EqVarDAG_HD}
\title{Title}
\usage{
EqVarDAG_HD(X, mtd = "ztest", alpha = 0.05, threshold = 0.1,
  FCD = NULL, precmtd = NULL)
}
\arguments{
\item{X}{- n x p data matrix}

\item{mtd}{-  methods for learning DAG from topological orderings.}

\item{alpha}{- desired selection significance level}

\item{threshold}{-  for rls and chol, the threshold level.}

\item{FCD}{- for debiased lasso, use the FCD procedure [False Discovery Rate Control via Debiased Lasso. Javanmard and Montanari. 2018]
or use individual tests to select support.}

\item{precmtd}{- for debiased lasso, how to compute debiasing matrix
"cv": node-wise lasso w/ joint 10 fold cv
"sqrtlasso": square-root lasso (no tune, default)}
}
\value{
- adjacency matrix and ordering
}
\description{
Title
}
