% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varEqChol.R
\name{get_DAGdata}
\alias{get_DAGdata}
\title{This function generates  (n xp) data. Borrowed from
EqVarDAG package}
\usage{
get_DAGdata(n, p, pc, type = "hub", err = "g", Bmin = 0.5)
}
\arguments{
\item{n}{- number of observations}

\item{p}{- number of features}

\item{pc}{- sparsity probability}

\item{type}{- type of the undirected graph}

\item{err}{- error distribution}
}
\value{
- X (n x p) data
}
\description{
This function generates  (n xp) data. Borrowed from
EqVarDAG package
}
