% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varEqChol.R
\name{get_ordering_ch}
\alias{get_ordering_ch}
\title{Find order from the Cholesky factor}
\usage{
get_ordering_ch(L, mat.T, p)
}
\arguments{
\item{L}{- Cholesky factor of precision matrix}

\item{mat.T}{- Cholesky factor of covariance matrix}

\item{p}{- dimensionality of the data}
}
\value{
ord - vector of ordering of variables
}
\description{
Find order from the Cholesky factor
}
