% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varEqChol.R
\name{rowdeletion}
\alias{rowdeletion}
\title{Implements rowdeletion algorithm as in Davis et.al (2005)}
\usage{
rowdeletion(L, k, const = 0.001)
}
\arguments{
\item{L}{- lower triangular Cholesky factor}

\item{k}{- index for row and column}

\item{const}{- constant value}
}
\value{
- Updated Cholesky factor
}
\description{
Implements rowdeletion algorithm as in Davis et.al (2005)
}
