% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varEqChol.R
\name{get_ordering_ch}
\alias{get_ordering_ch}
\title{Returns ordering from the Cholesky factor}
\usage{
get_ordering_ch(L, T, p)
}
\arguments{
\item{L}{- Cholesky factor of precision matrix}

\item{T}{- Cholesky factor of covariance matrix}

\item{p}{- dimensionality of the data}
}
\description{
Returns ordering from the Cholesky factor
}
